Add files via upload
Hello Everyone ! 
Myself, BUSHRA KAISER
Let's talk about logging in and signing up for my project in C# with DB connection:

Logging In: When we start the program, we see a login screen. We try to log in with random usernames and passwords. There's a checkbox to show the password if we need to double-check it. But when we click the login button, oops! We get an error message.

Signing Up: No worries, we can make a new account by clicking "sign up." We fill in our details like First name, Last name , email, and a username and password we'll remember. After hitting the register button, we get a message saying our account was created successfully.

Trying Again: Back at the login screen, we try logging in again with our new username and password. This time, when we hit login, success! We're in!

DB connection: All kinds of data for login and registration save to out created SQL server that already I ‘m connected my login registration forms.

                                          
Thank You Everyone